+++ packages/angular-workspace/nimble-angular/anchor-step/nimble-anchor-step-router-link-with-href.directive.ts
@@ -3,14 +3,6 @@ import { DisableableRouterLinkWithHrefDirective } from '@ni/nimble-angular';
 import type { AnchorStep } from './nimble-anchor-step.directive';
 
 /**
- * Selectors used for built-in Angular RouterLink directives:
- * RouterLink: ':not(a):not(area)[routerLink]'
- * RouterLinkWithHref: 'a[routerLink],area[routerLink]'
- *
- * See https://github.com/angular/angular/blob/5957ff4163f55d814be2cf80b9909244f1ce5262/packages/router/src/directives/router_link.ts
- */
-
-/**
  * Directive to handle nimble-anchor-step RouterLink support.
  * Note: Clients need to use [nimbleRouterLink] instead of [routerLink], so that there
  * won't also be an active RouterLink directive incorrectly handling navigation.