+++ packages/blazor-workspace/Tests/NimbleBlazor.Tests.Acceptance/Tests.InteractiveServer/ThemeProviderTests.cs
@@ -1,4 +1,3 @@
-﻿using BlazorWorkspace.Testing.Acceptance;
 using Microsoft.Playwright;
 using Xunit;
 
@@ -6,8 +5,8 @@ namespace NimbleBlazor.Tests.Acceptance.InteractiveServer;
 
 public class ThemeProviderTests : NimbleInteractiveAcceptanceTestsBase
 {
-    public ThemeProviderTests(PlaywrightFixture playwrightFixture, NimbleBlazorWebHostServerFixture blazorServerClassFixture)
-        : base(playwrightFixture, blazorServerClassFixture)
+    public ThemeProviderTests(NimbleBlazorWebHostServerFixture blazorServerClassFixture)
+        : base(blazorServerClassFixture)
     {
     }
 
@@ -23,12 +22,12 @@ public class ThemeProviderTests : NimbleInteractiveAcceptanceTestsBase
             var langIsInvalidCheckbox = page.Locator("nimble-checkbox", new PageLocatorOptions() { HasText = "InvalidLang" });
 
             await invalidButton.ClickAsync();
-            await Assertions.Expect(isValidCheckbox).Not.ToBeCheckedAsync();
-            await Assertions.Expect(langIsInvalidCheckbox).ToBeCheckedAsync();
+            await Expect(isValidCheckbox).Not.ToBeCheckedAsync();
+            await Expect(langIsInvalidCheckbox).ToBeCheckedAsync();
 
             await validButton.ClickAsync();
-            await Assertions.Expect(isValidCheckbox).ToBeCheckedAsync();
-            await Assertions.Expect(langIsInvalidCheckbox).Not.ToBeCheckedAsync();
+            await Expect(isValidCheckbox).ToBeCheckedAsync();
+            await Expect(langIsInvalidCheckbox).Not.ToBeCheckedAsync();
         }
     }
 }