+++ packages/angular-workspace/nimble-angular/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @ni/nimble-angular
 
-<!-- This log was last generated on Fri, 10 Jul 2026 18:20:51 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 11 Aug 2026 20:29:21 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 33.4.9
+
+Tue, 11 Aug 2026 20:29:21 GMT
+
+### Patches
+
+- Remove forked RouterLink implementation ([ni/nimble@e0da38d](https://github.com/ni/nimble/commit/e0da38d81bcd987bb2f4710e10bd34209cd4d3d1))
+
+## 33.4.8
+
+Tue, 04 Aug 2026 22:14:49 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.12.7
+
+## 33.4.7
+
+Wed, 29 Jul 2026 20:58:18 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.12.6
+
+## 33.4.6
+
+Wed, 22 Jul 2026 17:36:56 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.12.5
 
 ## 33.4.5