+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-components
 
-<!-- This log was last generated on Fri, 10 Jul 2026 18:20:51 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 04 Aug 2026 22:14:49 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 35.12.7
+
+Tue, 04 Aug 2026 22:14:49 GMT
+
+### Patches
+
+- Update to fast-foundation 10.2.5 ([ni/nimble@97dda99](https://github.com/ni/nimble/commit/97dda992014c79150846596631c19bdb00fdd406))
+
+## 35.12.6
+
+Wed, 29 Jul 2026 20:58:18 GMT
+
+### Patches
+
+- npm dependency updates ([ni/nimble@6c8854c](https://github.com/ni/nimble/commit/6c8854c79818d5d4b47d43ea5327661cb2a37f3c))
+
+## 35.12.5
+
+Wed, 22 Jul 2026 17:36:56 GMT
+
+### Patches
+
+- Forward autocomplete attribute to nimble-text-field internal input for password manager / autofill support ([ni/nimble@793419c](https://github.com/ni/nimble/commit/793419cf85a74d38242eb23bab23478c9004db0e))
 
 ## 35.12.4